Add files via upload
Class Bot changed
Edited new class ResQueue, containing logic of application and work with Queue and Resources
