+++ src/main/java/Bot.java
@@ -22,39 +22,11 @@ import static org.telegram.abilitybots.api.objects.Privacy.PUBLIC;
 
 
 public class Bot extends AbilityBot {
-    class MyTimerTask extends TimerTask {
-        final Task task;
 
-        MyTimerTask(Task task) {
-            this.task = task;
-        }
-
-        @Override
-        public void run() {
-            OnResource(task);
-        }
-    }
-
-    private int resourcesCount = BotConstants.STANDART_RESOURCES_COUNT;
-
-    private static Comparator<Task> maxComparator = new Comparator<Task>() {
-        @Override
-        public int compare(Task t1, Task t2) {
-            return t1.priority - t2.priority;
-        }
-    };
-    private static Comparator<Task> minComparator = new Comparator<Task>() {
-        @Override
-        public int compare(Task t1, Task t2) {
-            return t2.priority - t1.priority;
-        }
-    };
-    //is a threat sage and can't add null & not-comparable object
-    private PriorityBlockingQueue<Task> myQueue = new PriorityBlockingQueue<>(10, maxComparator);
-    private PriorityBlockingQueue<Task> myResources = new PriorityBlockingQueue<>(2, minComparator);
+    ResQueue resQueue = new ResQueue(this);
 
     @Nullable
-    private synchronized ChatMember getMember(long userId, long chatId) {
+    public synchronized ChatMember getMember(long userId, long chatId) {
         GetChatMember chatMember = new GetChatMember();
         chatMember.setUserId((int) userId);
         chatMember.setChatId(chatId);
@@ -67,7 +39,7 @@ public class Bot extends AbilityBot {
         return null;
     }
 
-    private synchronized void sendMsg(String chatId, String s) {
+    public synchronized void sendMsg(String chatId, String s) {
         SendMessage sendMessage = new SendMessage();
         sendMessage.enableMarkdown(true);
         sendMessage.setChatId(chatId);
@@ -79,96 +51,6 @@ public class Bot extends AbilityBot {
         }
     }
 
-    private void fromQueueToResource() {
-        Task task = myQueue.poll();
-        if (task == null) {
-            return;
-        }
-        myResources.add(task);
-        sendMsg(task.chatMember.getUser().getId().toString(), "your task on resource");
-        task.timer = new Timer();
-        task.timer.schedule(new MyTimerTask(task), BotConstants.timeOnResource);
-    }
-
-    private void OnResource(Task task) {
-        sendMsg(task.chatMember.getUser().getId().toString(), "your task finished");
-        myResources.remove(task);
-        fromQueueToResource();
-    }
-
-    @NotNull
-    private String getIn(MessageContext ctx) {
-        ChatMember chatMember = getMember(ctx.user().id(), ctx.chatId());
-        if (chatMember == null) {
-            return "no member";
-        }
-        int priority = Integer.parseInt(ctx.arguments()[0]);
-        Task task = new Task(chatMember, priority,BotConstants.timeOnResource);
-        return takeTask(task);
-    }
-
-    @NotNull
-    private String takeTask(Task task) {
-        if (task == null) {
-            return "no task";
-        }
-        if (myResources.size() < resourcesCount) {
-            myResources.add(task);
-            task.timer = new Timer();
-            task.timer.schedule(new MyTimerTask(task), BotConstants.timeOnResource);
-            return "You are on resource";
-        }
-        else if(task.priority > myResources.peek().priority){
-            return changeQR(task);
-        }
-        myQueue.add(task);
-        return "You are on queue";
-    }
-
-    private String getFirst(MessageContext ctx) {
-        if (myQueue.isEmpty())
-            return "Nobody";
-        else if (ctx.user().id() == myQueue.peek().chatMember.getUser().getId())
-            return "You are on, stupid monkey " + myQueue.peek().priority + " " + myQueue.peek().duration;
-        else
-            return myQueue.peek().chatMember.getUser().getUserName();
-    }
-
-    private String updateResources(MessageContext ctx) {
-        int newResourceCount = Integer.parseInt(ctx.arguments()[0]);
-        int dif = newResourceCount - resourcesCount;
-        resourcesCount = newResourceCount;
-        if (dif < 0) {
-            while (!myResources.isEmpty() && dif < 0) {
-                //returns null if this queue is empty.
-                Task task = myResources.poll();
-                sendMsg(task.chatMember.getUser().getId().toString(), takeTask(task));
-                ++dif;
-            }
-        } else if (dif > 0) {
-            while (!myQueue.isEmpty() && dif > 0) {
-                --dif;
-                fromQueueToResource();
-            }
-        }
-        return "Resources updated";
-    }
-
-    private String changeQR(Task task){
-        if(myResources.isEmpty()){
-            return "empty resources";
-        }
-
-        Task tmp = myResources.poll();
-        tmp.timer.cancel();
-        long dif = tmp.endTime.getEpochSecond();
-        dif -= Instant.now().getEpochSecond();
-        tmp.reSetEndTime(dif);
-        myQueue.add(tmp);
-        sendMsg(tmp.chatMember.getUser().getId().toString(), "your now is queue, your priority too low");
-        return takeTask(task);
-    }
-
     public Ability replyToGetInQ() {
         //todo check priority from 1-3
         return Ability
@@ -178,7 +60,7 @@ public class Bot extends AbilityBot {
                 .input(1)
                 .locality(USER)
                 .privacy(PUBLIC)
-                .action(ctx -> silent.send(getIn(ctx), ctx.chatId()))
+                .action(ctx -> silent.send(resQueue.getIn(ctx), ctx.chatId()))
                 .build();
     }
 
@@ -192,7 +74,7 @@ public class Bot extends AbilityBot {
                 .locality(USER)
                 .privacy(ADMIN)
                 .action(ctx -> {
-                    silent.send(updateResources(ctx), ctx.chatId());
+                    silent.send(resQueue.updateResources(ctx), ctx.chatId());
                 })
                 .build();
     }
@@ -204,7 +86,7 @@ public class Bot extends AbilityBot {
                 .info("Tell who the first on queue now")
                 .locality(USER)
                 .privacy(PUBLIC)
-                .action(ctx -> silent.send(getFirst(ctx), ctx.chatId()))
+                .action(ctx -> silent.send(resQueue.getFirst(ctx), ctx.chatId()))
                 .build();
     }
